requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
motor==3.6.0
python-dotenv==1.0.1
pydantic==2.9.2
email_validator==2.2.0
Jinja2==3.1.4
python_jose==3.3.0
websockets==14.1
